pyproject.toml
[project]
name = "neuromansui"
version = "0.1.0"
description = ""
authors = [
    {name = "Kevin Zeidler",email = "kevin.zeidler@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
packages = [{ include = "neuromansui", from = "." }]

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
langchain = "^0.3.19"
openai = "^1.64.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
